plugins/pensive/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pensive"
version = "1.9.4"
authors = [{name = "Alex Thola", email = "alexthola@gmail.com"}]
description = "Reflective code review plugin with specialized analysis skills"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "psutil>=5.8.0",
]

[project.urls]
Homepage = "https://github.com/anthropics/pensive"
Documentation = "https://pensive.readthedocs.io/"
Repository = "https://github.com/anthropics/pensive.git"
"Bug Tracker" = "https://github.com/anthropics/pensive/issues"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.0",
    "coverage>=7.0.0",
    "ruff>=0.14.13",
    "mypy>=1.11.0",
    "pre-commit>=4.0.0",
    "types-PyYAML>=6.0.0",
]

[tool.setuptools_scm]
write_to = "src/pensive/_version.py"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=pensive",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=90",
    "--tb=short",
]
testpaths = ["tests"]
pythonpath = [".", "src"]
markers = [
    "unit: Unit tests for individual functions and methods",
    "integration: Integration tests for workflows and end-to-end scenarios",
    "slow: Tests that take longer than 1 second to run",
    "security: Security-focused tests and vulnerability detection",
    "architecture: Architecture and design pattern tests",
    "performance: Performance and scalability tests",
    "edge_case: Edge case and boundary condition tests",
    "bdd: BDD-style behavior-driven tests with Given/When/Then structure",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/pensive"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "setup.py",
    "src/pensive/_version.py",
    "src/pensive/skills/rust_review_data.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "@overload",
]
fail_under = 90
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
extend-ignore = ["S101"]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL",  # pylint
    "RUF", # Ruff-specific rules
]
ignore = [
    "B008",   # do not perform function calls in argument defaults
]
unfixable = [
    "B006",   # Do not use mutable data structures for argument defaults
    "RUF001", # Remove ambiguous unicode character
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = [
    "ARG001",  # Unused function args in tests
    "ARG002",  # Unused method args in tests
    "PLR2004", # Magic value used in comparison
    "S101",    # Use of assert detected
    "D102",    # Missing method docstrings in test classes
    "D103",    # Missing function docstrings in tests
    "S108",    # Allow hardcoded temp paths in tests
    "D205",    # Multiline docstring summary (Given/When/Then patterns)
]
"__init__.py" = [
    "F401",   # Unused import
]
"src/pensive/skills/makefile_review.py" = [
    "SIM102",  # Complex nested checks in analysis logic
]
"src/pensive/skills/architecture_review.py" = [
    "SIM102",  # Complex nested checks in analysis logic
]
"src/pensive/skills/tiered_audit.py" = [
    "PLR2004", # Magic value in path splitting logic
]
"src/pensive/reporting/formatters.py" = [
    "ARG002", # repo_path kept for API consistency
]
"src/pensive/analysis/repository_analyzer.py" = [
    "SIM102", # Complex conditional logic in analysis
]
"src/pensive/integrations/cicd.py" = [
    "SIM102", # Complex conditional logic in CI/CD detection
    "ARG002", # repo_path kept for API consistency
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "psutil.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["leyline.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["gauntlet.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pensive.skills.rust_review.cargo",
    "pensive.skills.rust_review.builtins",
    "pensive.skills.rust_review.ownership",
    "pensive.skills.rust_review.patterns",
    "pensive.skills.rust_review.safety",
    "pensive.skills.rust_review.structure",
]
disable_error_code = ["attr-defined"]
